Added new delegate method: PlantingEstablish to ensure that establishment is > 0.0 when planting occurs.  All other changes cosmetic.
git-svn-id: http://Marc-PC/svn/Full@749 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/succession-library/trunk/src/initial-communities/DatasetParser.cs
@@ -37,11 +37,11 @@ namespace Landis.Library.InitialCommunities
             
             Dataset dataset = new Dataset();
 
-            InputVar<ushort> mapCode = new InputVar<ushort>("MapCode");
+            InputVar<uint> mapCode = new InputVar<uint>("MapCode");
             InputVar<string> speciesName = new InputVar<string>("Species");
             InputVar<ushort> age = new InputVar<ushort>("Age");
 
-            Dictionary <ushort, int> mapCodeLineNumbers = new Dictionary<ushort, int>();
+            Dictionary <uint, int> mapCodeLineNumbers = new Dictionary<uint, int>();
 
             while (! AtEndOfInput) {
                 //  Read initial community
